adds pundit to remaining controllers
  accounts
  bills
  categories
  customers
  invoices
  reports
  settings
  suppliers

+++ app/controllers/customers_controller.rb
@@ -1,16 +1,20 @@
 class CustomersController < ApplicationController
   before_action :authenticate_user!
+  after_action :verify_authorized
 
   def index
     @customers = current_account.customers if current_account
+    authorize @customers
   end
 
   def new
     @customer = current_account.customers.build
+    authorize @customer
   end
 
   def create
     @customer = Customer.new(cust_params)
+    authorize @customer
     if @customer.save
       redirect_to customers_path, notice: "Customer successfully created"
     else
@@ -20,10 +24,12 @@ class CustomersController < ApplicationController
 
   def edit
     @customer = current_account.customer(params[:id])
+    authorize @customer
   end
 
   def update
     @customer = current_account.customer(params[:id])
+    authorize @customer
     if @customer.update_attributes(cust_params)
       redirect_to customers_url, notice: "Customer successfully updated"
     else
@@ -33,6 +39,7 @@ class CustomersController < ApplicationController
 
   def destroy
     @customer = current_account.customer(params[:id])
+    authorize @customer
     if @customer.destroy
       msg = "#{@customer.name} destroyed"
     else
@@ -44,6 +51,6 @@ class CustomersController < ApplicationController
   private
 
   def cust_params
-    params.require(:customer).permit(:account_id, :name, :address, :postcode, :is_default)
+    params.require(:customer).permit(*policy(@customer || Customer).permitted_attributes)
   end
 end
